requirements.txt
langchain
streamlit
langchain_groq
langchain_community
python-dotenv
wikipedia
numexpr